libpayload/libc/time: Fix possible overflow in multiplication
The value from raw_read_cntfrq_el0() could be large enough to cause
overflow when multiplied by USECS_PER_SEC. To prevent this, both
USECS_PER_SEC and hz can be reduced by dividing them by their GCD.

This patch also modifies the return type of `timer_hz()` from
`uint64_t` to `uint32_t`, assuming that in practice the timestamp
counter should never be that fast.

BUG=b:307790895
TEST=boot to kernel and check the timestamps from `cbmem`

Reviewed-on: https://review.coreboot.org/c/coreboot/+/78888
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ payloads/libpayload/drivers/timer/arm64_arch_timer.c
@@ -32,7 +32,7 @@
 #include <arch/lib_helpers.h>
 #include <libpayload.h>
 
-uint64_t timer_hz(void)
+uint32_t timer_hz(void)
 {
 	return raw_read_cntfrq_el0();
 }
