+++ payloads/libpayload/drivers/timer/generic.c
@@ -33,7 +33,7 @@
 #include <assert.h>
 #include <libpayload.h>
 
-uint64_t timer_hz(void)
+uint32_t timer_hz(void)
 {
 	/* libc/time.c currently requires all timers to be at least 1MHz. */
 	assert(CONFIG_LP_TIMER_GENERIC_HZ >= 1000000);
